backend/requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-google-genai
google-generativeai
tavily-python
wikipedia 
python-dotenv
streamlit 
requests